+++ ShopperInsights/src/main/java/com/braintreepayments/api/shopperinsights/v2/CustomerRecommendationsResult.kt
@@ -1,0 +1,25 @@
+package com.braintreepayments.api.shopperinsights.v2
+
+import com.braintreepayments.api.core.ExperimentalBetaApi
+
+/**
+ * Represents the result of customer payment option recommendations generation.
+ */
+@ExperimentalBetaApi
+sealed class CustomerRecommendationsResult {
+    /**
+     * Indicates a successful customer recommendations generation.
+     *
+     * @property customerRecommendations The customer recommendations generated.
+     */
+    class Success internal constructor(
+        val customerRecommendations: CustomerRecommendations
+    ) : CustomerRecommendationsResult()
+
+    /**
+     * Indicates a failure during customer recommendations generation.
+     *
+     * @property error The exception that caused the failure.
+     */
+    class Failure internal constructor(val error: Exception) : CustomerRecommendationsResult()
+}